part2_03: fix typos, tweak formatting, increase PESTPP-IES runs
The first part of the tutorial uses PESTPP-SWP to complete 250
forward model runs. The second part of the tutorial uses PESPP-IES
but was only configured to perform 50 model runs. The text said
the second round of MC should take about the same time, so I
increased the PEST-IES ensemble count to 250.

I also capitalized the names of all PEST++ programs in the tutorial
text to match the formatting of the PEST++ manual.

I removed a reference in the text saying to look at the terminal
window for PEST++ output. In my vanilla installation of Jupyter
Lab from Anaconda, all PEST++ output has been appearing in the
notebook itself.